refactor(internal/blocktest): add package for shared test code #27270
Move the testHasher helper to internal/blocktest/test_hash.go and update all usages to import from there. This reduces code duplication and improves maintainability.
